Curators component
